Fix Honkai and Genshin Game Settings file in the wrong place
