+++ CollapseLauncher/Classes/GameManagement/GameSettings/Honkai/Settings.cs
@@ -6,15 +6,17 @@ using Microsoft.Win32;
 using System.IO;
 using static CollapseLauncher.GameSettings.Statics;
 
-namespace CollapseLauncher.GameSettings.Genshin
+namespace CollapseLauncher.GameSettings.Honkai
 {
-    internal class GenshinSettings : ImportExportBase, IGameSettings, IGameSettingsUniversal
+    internal class HonkaiSettings : ImportExportBase, IGameSettings, IGameSettingsUniversal
     {
         public CustomArgs SettingsCustomArgument { get; set; }
+        public PersonalGraphicsSettingV2 SettingsGraphics { get; set; }
+        public PersonalAudioSetting SettingsAudio { get; set; }
         public BaseScreenSettingData SettingsScreen { get; set; }
         public CollapseScreenSetting SettingsCollapseScreen { get; set; }
 
-        public GenshinSettings(PresetConfigV2 gameConfig)
+        public HonkaiSettings(PresetConfigV2 gameConfig)
         {
             // Init Root Registry Key
             RegistryPath = Path.Combine(RegistryRootPath, gameConfig.InternalGameNameInConfig);
@@ -33,7 +35,11 @@ namespace CollapseLauncher.GameSettings.Genshin
         private void InitializeSettings()
         {
             // Load Settings
+            SettingsGraphics = PersonalGraphicsSettingV2.Load();
+            SettingsAudio = PersonalAudioSetting.Load();
             SettingsCustomArgument = CustomArgs.Load();
+            SettingsScreen = ScreenSettingData.Load();
+            SettingsCollapseScreen = CollapseScreenSetting.Load();
         }
 
         public void RevertSettings() => InitializeSettings();
@@ -41,7 +47,11 @@ namespace CollapseLauncher.GameSettings.Genshin
         public void SaveSettings()
         {
             // Save Settings
+            SettingsGraphics.Save();
+            SettingsAudio.Save();
             SettingsCustomArgument.Save();
+            SettingsScreen.Save();
+            SettingsCollapseScreen.Save();
         }
 
         public IGameSettingsUniversal AsIGameSettingsUniversal() => this;
